Se añade LOG al problema del pete relacionado con el fault a la hora de acceder a una categoría calculando el balance de los años Se soluciona el problema de las fuentes helvéticas Se añade nueva versión de Flurry Se corrigen múltiples warnins que han salido con la nueva versión de xCode relativos a funciones no utilizadas

+++ Code/IncomesAndExpenses/IAEDayCalendarSelectorViewController.m
@@ -83,7 +83,7 @@ static const CGFloat kWhiteAlphaComponentForCircle = 1.0;
 
 - (void)configureNavigationBarTitle
 {
-    NSString *yearName = [NSString stringWithFormat:@"%d", self.yearDate];
+    NSString *yearName = [NSString stringWithFormat:@"%lu", (unsigned long)self.yearDate];
     NSString *monthName = [IAEDateHelper findMonthNameStringWithMonthIndex:self.monthIndex inShortForm:NO];
 
     self.navigationBarTitle.title = [NSString stringWithFormat:NSLocalizedString(@"LTEXT_CALENDARDAYSELECTOR_TITLE", ""), monthName, yearName];
@@ -131,7 +131,7 @@ static const CGFloat kWhiteAlphaComponentForCircle = 1.0;
                                               rowIt * self.dayOfTheMonthHeightSize,
                                               self.dayOfTheMonthWidthSize,
                                               self.dayOfTheMonthHeightSize);
-                NSString *labelText = [NSString stringWithFormat:@"%d", dayIt];
+                NSString *labelText = [NSString stringWithFormat:@"%lu", (unsigned long)dayIt];
                 NSDictionary *labelAttributes = @{NSFontAttributeName: [UIFont fontWithName:kDayOfTheMonthFontFamilyName size:kDayOfTheMonthFontSize],
                                                   NSForegroundColorAttributeName: [UIColor darkTextColor]};
