+++ Code/IncomesAndExpenses/IAEHelpIndexViewController.m
@@ -132,9 +132,9 @@ static const NSUInteger kRowOfAboutIndex = 2;
 - (void)configureCell:(UITableViewCell *)cell ofTableView:(UITableView *)tableView forIndexPath:(NSIndexPath *)indexPath
 {
     const NSUInteger indexSufix = indexPath.row + 1;
-    NSString *ltext = [NSString stringWithFormat:@"LTEXT_SETTINGSINDEX_%d", indexSufix];
+    NSString *ltext = [NSString stringWithFormat:@"LTEXT_SETTINGSINDEX_%lu", (unsigned long)indexSufix];
     cell.textLabel.text = NSLocalizedString(ltext, @"");
-    NSString *imageName = [NSString stringWithFormat:@"settingsindex_%d", indexSufix];
+    NSString *imageName = [NSString stringWithFormat:@"settingsindex_%lu", (unsigned long)indexSufix];
     cell.imageView.image = [UIImage imageNamed:imageName];
     cell.accessoryType = UITableViewCellAccessoryDisclosureIndicator;
 }
